Move EditedRenderer
